pyproject.toml
[project]
name = "mcpintegrationcrewai"
version = "0.1.0"
description = "Building your own Custom MCP Servers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.6.0",
    "requests>=2.32.3",
    "crewai-tools[mcp]",
    "crewai",
    "agentops",
    "python-dotenv",
]